реорганизация файлов проекта
