+++ users/sivanov/lessons/lesson_31_sql_hw2/stuff.py
@@ -1,5 +1,5 @@
 """
-модуль, содержащий функционал работы с базой данных для бухгалтерии
+модуль, содержащий функционал работы с базой данных
 """
 import sqlite3
